Resolves a merge conflict
Rebase. This commit includes the following commits:

feat(various): Installed the vite-plugin-pwa package; Added the PWA plugin to the vite.config.js file

feat(various): Work in progress

fix(vite.config): Fixed a bug that was related to the UserConfigExport type

feat(various): The service worker is now being registered.

feat(various): Work in Progress (cont)

fix(various): Registering the service worker is no longer a problem

refactor(various): Deleted the original code; Organized the code a bit;
Tried the injectManifest approach. However, the service worker isn't
being registered at the moment.

feat(various): Finally figured out the cors issue

feat(offline and App): Removed the old service worker registration code
and the offline.html file

feat(various): Removed the dev-dist folder and the devOptions property
from the vite config file
