+++ index.html
@@ -7,7 +7,6 @@
     <meta name="theme-color" content="#000000" />
     <meta name="React Boilerplate" content="BWTC React Boilerplate App" />
     <link rel="apple-touch-icon" href="/logo192.png" />
-    <link rel="manifest" href="/manifest.json" />
     <title>React Boilerplate</title>
   </head>
   <body>
