+++ package.json
@@ -95,6 +95,7 @@
     "typescript-plugin-styled-components": "^2.0.0",
     "vite": "^4.1.1",
     "vite-plugin-node-polyfills": "^0.7.0",
+    "vite-plugin-pwa": "^0.14.7",
     "vite-tsconfig-paths": "^4.0.5"
   },
   "lint-staged": {
