--- vite.config.js
@@ -1,28 +1,0 @@
-import { defineConfig, loadEnv } from 'vite';
-import react from '@vitejs/plugin-react';
-import tsconfigPaths from 'vite-tsconfig-paths';
-import { nodePolyfills } from 'vite-plugin-node-polyfills';
-import dns from 'dns';
-
-dns.setDefaultResultOrder('verbatim');
-
-// https://vitejs.dev/config/
-export default defineConfig(({ mode }) => {
-  const env = loadEnv(mode, process.cwd(), '');
-
-  return {
-    plugins: [
-      react(),
-      tsconfigPaths(),
-      nodePolyfills({
-        protocolImports: true,
-      }),
-    ],
-    server: {
-      port: process.env.PORT || 4200,
-    },
-    define: {
-      'process.env': env,
-    },
-  };
-});
